+++ front/src/components/DocumentScores.js
@@ -15,11 +15,11 @@ import { Tooltip as InfoTooltip } from "react-tooltip";
 
 import styles from "@/styles/ReportScores.module.css";
 
-const CustomTooltip = ({ active, payload, label, measurement }) => {
+const CustomTooltip = ({ active, payload, label, measurement, total }) => {
   if (active && payload && payload.length) {
     return (
       <div className={styles.tooltip}>
-        <p className={styles.label}>{label}</p>
+        <p className={styles.label}>Score {label}</p>
         {payload.map((item, idx) => {
           return (
             <>
@@ -27,14 +27,7 @@ const CustomTooltip = ({ active, payload, label, measurement }) => {
                 key={idx}
                 className={styles.desc}
                 style={{ color: item.fill }}
-              >{`${measurement}:   ${numberFormatter(item.value)}`}</p>
-              {item.payload.perc && (
-                <p
-                  key={idx}
-                  className={styles.perc}
-                  style={{ color: item.fill }}
-                >{`% of total:   ${item.payload.perc} %`}</p>
-              )}
+              >{`${numberFormatter(item.value)} ${measurement.toLowerCase()}`} <span style={{ fontWeight: 600 }}>{`(${((item.value) / total * 100).toFixed(2)}%)`}</span></p>
             </>
           );
         })}
@@ -97,9 +90,9 @@ export default function DocumentScores({ scores, footNote }) {
         <InfoTooltip
           anchorSelect=".doc-scores-distribution-info"
           place="top"
-          clickable
+          clickable style={{ fontWeight: 400, backgroundColor: "rgba(17, 21, 24, 1)", zIndex: 10000 }}
         >
-          <div style={{ display: "flex", flexDirection: "column" }}>
+          <div style={{ display: "flex", flexDirection: "column", fontSize: "14px" }}>
             <span>
               Obtained with Web Docs Scorer (
               <a
@@ -122,7 +115,7 @@ export default function DocumentScores({ scores, footNote }) {
               {numberFormatter(underFive)} documents)
             </p>
             <p>
-              score {">="} 5 - <strong>{+percOverFive.toFixed(2)}%</strong> (
+              score {"≥"} 5 - <strong>{+percOverFive.toFixed(2)}%</strong> (
               {numberFormatter(overFive)} documents)
             </p>
           </div>
@@ -133,7 +126,7 @@ export default function DocumentScores({ scores, footNote }) {
             data={processedItems}
             margin={{
               top: 32,
-              right: 20,
+              right: 0,
               left: 10,
               bottom: 25,
             }}
@@ -147,6 +140,7 @@ export default function DocumentScores({ scores, footNote }) {
               allowDecimals
               domain={[0, 10]}
               ticks={docScoresNums}
+              padding={{ left: 20, right: 20 }}
             >
               <Label
                 value="Scores"
@@ -168,7 +162,7 @@ export default function DocumentScores({ scores, footNote }) {
               tickFormatter={DataFormatter}
             />
             <Tooltip
-              content={<CustomTooltip measurement={"Documents"} />}
+              content={<CustomTooltip measurement={"Documents"} total={totalValue} />}
               wrapperStyle={{ outline: "none" }}
             />
             <ReferenceLine y={0} stroke="#000" />