+++ front/src/components/DocumentSizes.js
@@ -15,28 +15,21 @@ import {
   ResponsiveContainer,
 } from "recharts";
 
-import { DataFormatter } from "@/lib/helpers";
+import { DataFormatter, numberFormatter } from "@/lib/helpers";
 
-const CustomTooltip = ({ active, payload, label, measurement }) => {
+const CustomTooltip = ({ active, payload, label, total }) => {
   if (active && payload && payload.length) {
     return (
       <div className={styles.tooltip}>
-        <p className={styles.label}>{label}</p>
+        <p className={styles.label}>{label} Segments</p>
         {payload.map((item, idx) => {
           return (
             <>
               <p
                 key={idx}
                 className={styles.desc}
-                style={{ color: item.fill }}
-              >{`${measurement}:   ${Intl.NumberFormat("en", {
-                notation: "compact",
-              }).format(item.value)}`}</p>
-              {payload[0].payload.perc && (
-                <p
-                  className={styles.perc}
-                >{`% of total:   ${payload[0].payload.perc} %`}</p>
-              )}
+              >{`${numberFormatter(item.value)} documents`} <span style={{ fontWeight: 600 }}>{`(${((item.value / total) * 100).toFixed(2)}%)`}</span></p>
+
             </>
           );
         })}
@@ -55,6 +48,7 @@ function NewDocumentSizes({ documentSizesObj }) {
     remainingSum,
     percentageOfTotal,
     remainingPercentage,
+    total
   } = documentSizesObj;
 
   return (
@@ -75,32 +69,28 @@ function NewDocumentSizes({ documentSizesObj }) {
                 />
               )}
             </a>
-            <InfoTooltip anchorSelect=".segments-info-graph" place="top">
-              Segments correspond to paragraph and list boundaries as
-              defined by HTML elements{" "}
-              <code>
-                ({"<"}p{">"}, {"<"}ul{">"}, {"<"}ol{">"}, etc.)
-              </code>{" "}
-              replaced by newlines.
+            <InfoTooltip anchorSelect=".segments-info-graph" place="top" style={{ fontWeight: 400, backgroundColor: "rgba(17, 21, 24, 1)", zIndex: 10000 }}>
+              <p style={{ fontSize: "14px" }}> Segments correspond to paragraph and list boundaries as
+                defined by HTML elements{" "}
+                <code>
+                  ({"<"}p{">"}, {"<"}ul{">"}, {"<"}ol{">"}, etc.)
+                </code>{" "}
+                replaced by newlines.</p>
             </InfoTooltip>
           </div>
           <div className={styles.documentSizes}>
             {percentageOfTotal && remainingSum >= 0 && (
               <div className={styles.reportTitleDocsSizes}>
                 <p>
-                  <strong>{"<="} 25</strong> segments{" "}
+                  <strong>{"≤"} 25</strong> segments{" "}
                   <strong>{+percentageOfTotal.toFixed(2)}%</strong> (
-                  {Intl.NumberFormat("en", {
-                    notation: "compact",
-                  }).format(totalDocuments)}{" "}
+                  {numberFormatter(totalDocuments)}{" "}
                   documents)
                 </p>
                 <p>
                   <strong>{">"} 25</strong> segments{" "}
                   <strong>{+remainingPercentage.toFixed(2)}%</strong> (
-                  {Intl.NumberFormat("en", {
-                    notation: "compact",
-                  }).format(remainingSum)}{" "}
+                  {numberFormatter(remainingSum)}{" "}
                   documents)
                 </p>
               </div>
@@ -125,6 +115,7 @@ function NewDocumentSizes({ documentSizesObj }) {
                   tickMargin={5}
                   domain={[0, 25]}
                   ticks={numbers}
+                  padding={{ left: 10, right: 15 }}
                 >
                   {" "}
                   <Label
@@ -146,25 +137,24 @@ function NewDocumentSizes({ documentSizesObj }) {
                   tickFormatter={DataFormatter}
                 />
                 <Tooltip
-                  content={<CustomTooltip measurement="Documents" />}
+                  content={<CustomTooltip measurement="Documents" total={total} />}
                   wrapperStyle={{ outline: "none" }}
                 />
                 <ReferenceLine y={0} stroke="#000" />
                 <Bar
                   dataKey={(val) => val[1]}
-                  fill="#38686a"
+                  fill="#38686aff"
                   maxBarSize={30}
                 >
                   {" "}
                   <LabelList
                     dataKey={(val) =>
-                      Intl.NumberFormat("en", {
-                        notation: "compact",
-                      }).format(val[1])
+                      numberFormatter(val[1])
                     }
+                    fill="#244446ff"
                     position="top"
-                    fontWeight={600}
-                    fontSize={10}
+                    fontWeight={800}
+                    fontSize={11}
                   />
                 </Bar>
               </BarChart>