+++ front/src/components/LangDocs.js
@@ -8,28 +8,25 @@ import {
   ResponsiveContainer,
   LabelList,
 } from "recharts";
-import InfoCircle from "./InfoCircle";
-import { Tooltip as InfoTooltip } from "react-tooltip";
-import {
-  DataFormatter,
-  numberFormatter,
-  percFormatter,
-} from "@/lib/helpers";
+import { DataFormatter, numberFormatter, percFormatter } from "@/lib/helpers";
 
 import styles from "@/styles/LangDocs.module.css";
 
-const CustomTooltip = ({ active, payload, label, srclang }) => {
+const CustomTooltip = ({ active, payload, label, srclang, total }) => {
   if (active && payload && payload.length) {
     return (
       <div className={styles.tooltip}>
-        <p className={styles.label}>{`${label}% of segments in ${srclang}`}</p>
+        <p
+          className={styles.label}
+        >{`${label}% of segments in ${srclang[0].label}`}</p>
         {payload.map((item, idx) => {
           return (
-            <p
-              key={idx}
-              className={styles.desc}
-              style={{ color: item.fill }}
-            >{`Documents: ${item.payload[1].toLocaleString("en-US")} `}</p>
+            <p key={idx} className={styles.desc} style={{ color: item.fill }}>
+              {`${item.payload[1].toLocaleString("en-US")} documents`}{" "}
+              <span style={{ fontWeight: 600 }}>
+                ({((item.payload[1] / total) * 100).toFixed(2)}%)
+              </span>
+            </p>
           );
         })}
       </div>
@@ -39,17 +36,42 @@ const CustomTooltip = ({ active, payload, label, srclang }) => {
   return null;
 };
 
-function LangDocs({ langDocs, srclang, footNote }) {
+function LangDocs({ langDocs, srclang }) {
   const numbers = [0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100];
+  const { total, totalUnderFifty, totalOverFifty } = langDocs.reduce(
+    (acc, cur) => {
+      acc.total += cur[1];
+
+      if (cur[0] < 0.5) {
+        acc.totalUnderFifty += cur[1];
+      }
+      if (cur[0] >= 0.5) {
+        acc.totalOverFifty += cur[1];
+      }
+      return acc;
+    },
+    { total: 0, totalUnderFifty: 0, totalOverFifty: 0 }
+  );
+  const underFifty = ((totalUnderFifty / total) * 100).toFixed(2);
+  const overFifty = ((totalOverFifty / total) * 100).toFixed(2);
   return (
     <div className={styles.langDocs}>
-      <ResponsiveContainer width="100%" height="100%">
+      <div className={styles.stats}>
+        <p style={{ marginBottom: "6px" }}>
+          segments {"<"} 50% - <span>{underFifty}% </span>(
+          {numberFormatter(totalUnderFifty)} documents)
+        </p>
+        <p>
+          segments {"≥"} 50% - <span>{overFifty}%</span> (
+          {numberFormatter(totalOverFifty)} documents)
+        </p>
+      </div>
+      <ResponsiveContainer width="100%" height="85%">
         <BarChart
-          height={300}
           data={langDocs}
           margin={{
             top: 32,
-            right: 20,
+            right: 10,
             left: 10,
             bottom: 15,
           }}
@@ -63,7 +85,7 @@ function LangDocs({ langDocs, srclang, footNote }) {
               angle: 0,
               position: "bottom",
               offset: 0,
-              fontSize: 14,
+              fontSize: 12,
             }}
             ticks={numbers}
             type="number"
@@ -77,16 +99,16 @@ function LangDocs({ langDocs, srclang, footNote }) {
               value: "Documents",
               angle: 0,
               position: "top",
-              offset: 12,
+              offset: 18,
               fontSize: 14,
             }}
             fontSize={12}
           />
           <Tooltip
-            content={<CustomTooltip srclang={srclang} />}
+            content={<CustomTooltip srclang={srclang} total={total} />}
             wrapperStyle={{ outline: "none" }}
           />
-          <Bar dataKey={(value) => value[1]} fill="#6d466b" maxBarSize={50}>
+          <Bar dataKey={(value) => value[1]} fill="#582156ff" maxBarSize={50}>
             <LabelList
               dataKey={(value) => numberFormatter(value[1])}
               fill="#6d466b"