+++ front/src/components/Report.js
@@ -88,9 +88,11 @@ export default function Report({ date, report }) {
     setLoadingPdf(false);
   };
 
-  const docsDomains = report.docs_top100_domains?.slice(0, 10);
+  const docsDomains = report.docs_top100_domains?.length ? report.docs_top100_domains.slice(0, 10) : undefined;
 
-  const docsTLDs = report.docs_top100_tld?.slice(0, 10);
+  const docsTLDs = report.docs_top100_tld?.length
+    ? report.docs_top100_tld.slice(0, 10)
+    : undefined;
 
   const docs_total = report.docs_total;
 
@@ -211,7 +213,7 @@ export default function Report({ date, report }) {
                   </p>
                 </div>
               </div>
-              <div className={styles.volumes}>
+              <div className={styles.volumes} style={{ marginTop: "20px" }}>
                 <h3>Volumes</h3>
                 <table>
                   <thead>
@@ -220,16 +222,16 @@ export default function Report({ date, report }) {
                       <th>
                         <div className={styles.containsTooltip}>
                           Segments{" "}
-                          <a className="segments-info">
-                            {!footNote && <InfoCircle />}
+                          <a className="segments-info" >
+                            {!footNote && <InfoCircle style={{ marginBottom: "-2px", marginLeft: "3px" }} />}
                           </a>
-                          <Tooltip anchorSelect=".segments-info" place="top">
-                            Segments correspond to paragraph and list boundaries
-                            as defined by HTML elements{" "}
-                            <code>
-                              ({"<"}p{">"}, {"<"}ul{">"}, {"<"}ol{">"}, etc.)
-                            </code>{" "}
-                            replaced by newlines.
+                          <Tooltip anchorSelect=".segments-info" place="top" style={{ fontWeight: 400, backgroundColor: "rgba(17, 21, 24, 1)", zIndex: 10000 }}>
+                            <p style={{ fontSize: "14px" }}> Segments correspond to paragraph and list boundaries
+                              as defined by HTML elements{" "}
+                              <code>
+                                ({"<"}p{">"}, {"<"}ul{">"}, {"<"}ol{">"}, etc.)
+                              </code>{" "}
+                              replaced by newlines.</p>
                           </Tooltip>
                         </div>
                       </th>
@@ -239,31 +241,31 @@ export default function Report({ date, report }) {
                       {duplicationRatio !== "" && (
                         <th className={styles.desktopData}>Duplication ratio</th>
                       )}
-                      {!trglang && srcTokens && (
+                      {!trglang && Number(src_tokens) > 0 && (
                         <th className={styles.desktopData}>
                           <div className={styles.containsTooltip}>
                             Tokens{" "}
                             <a className="tokens-info">
-                              {!footNote && <InfoCircle />}
+                              {!footNote && <InfoCircle style={{ marginBottom: "-2px", marginLeft: "3px" }} />}
                             </a>
                             <Tooltip
                               anchorSelect=".tokens-info"
                               place="top"
-                              clickable
+                              clickable style={{ fontWeight: 400, backgroundColor: "rgba(17, 21, 24, 1)", zIndex: 100000 }}
                             >
-                              Tokenized with{" "}
-                              <a
-                                href="https://github.com/hplt-project/data-analytics-tool/blob/main/tokenizers-info.md"
-                                target="_blank"
-                                className={styles.tooltipLink}
-                              >
-                                https://github.com/hplt-project/data-analytics-tool/blob/main/tokenizers-info.md
-                              </a>
+                              <p style={{ fontSize: "14px" }}>  Tokenized with{" "}
+                                <a
+                                  href="https://github.com/hplt-project/data-analytics-tool/blob/main/tokenizers-info.md"
+                                  target="_blank"
+                                  className={styles.tooltipLink}
+                                >
+                                  https://github.com/hplt-project/data-analytics-tool/blob/main/tokenizers-info.md
+                                </a></p>
                             </Tooltip>
                           </div>
                         </th>
                       )}
-                      {trglang && srcTokens && (
+                      {trglang && Number(src_tokens) > 0 && (
                         <th className={styles.desktopData}>SL tokens</th>
                       )}
                       {trglang && srcChars && (
@@ -329,11 +331,11 @@ export default function Report({ date, report }) {
                       {duplicationRatio !== "" && (
                         <td className={styles.desktopData}>
                           {(duplicationRatio * 100).toFixed(2)}%
-                        </td>
-                      )}
-                      {srcTokens && (
-                        <td className={styles.desktopData}>{srcTokens}</td>
-                      )}
+                        </td>)}
+                      {Number(src_tokens) > 0 &&
+                        (<td className={styles.desktopData}>{srcTokens}</td>
+                        )}
+
 
                       {srcChars && (
                         <td className={styles.desktopData}>{srcChars}</td>
@@ -474,28 +476,29 @@ export default function Report({ date, report }) {
 
                     </h3>
                     <>
-                      <a className="lang-distribution-info">
+                      <a className="lang-distribution-info" style={{ marginLeft: "5px" }}>
                         {!footNote && <InfoCircle />}
                       </a>
                       <Tooltip
                         anchorSelect=".lang-distribution-info"
                         place="top"
                         clickable
+                        style={{ fontWeight: 400, backgroundColor: "rgba(17, 21, 24, 1)", zIndex: 10000 }}
                       >
-                        Language identified with FastSpell (
-                        <a
-                          href="https://github.com/mbanon/fastspell"
-                          target="_blank"
-                          className={styles.tooltipLink}
-                        >
-                          https://github.com/mbanon/fastspell
-                        </a>
-                        )
+                        <p style={{ fontSize: "14px" }}>   Language identified with FastSpell (
+                          <a
+                            href="https://github.com/mbanon/fastspell"
+                            target="_blank"
+                            className={styles.tooltipLink}
+                          >
+                            https://github.com/mbanon/fastspell
+                          </a>
+                          )</p>
                       </Tooltip>
                     </>
                   </div>
                 ) : (
-                  <h3>Source</h3>
+                  <h3 className={styles.smaller}>Source</h3>
                 )}
                 <LanguagePieChart
                   langs={report.src_langs}
@@ -515,7 +518,7 @@ export default function Report({ date, report }) {
 
                   </h3>
                   <>
-                    <a className="lang-distribution-info-second">
+                    <a className="lang-distribution-info-second" style={{ marginLeft: "5px" }}>
                       {!footNote && (
                         <InfoCircle
                         />
@@ -525,16 +528,17 @@ export default function Report({ date, report }) {
                       anchorSelect=".lang-distribution-info-second"
                       place="top"
                       clickable
+                      style={{ fontWeight: 400, backgroundColor: "rgba(24, 18, 17, 1)", zIndex: 10000 }}
                     >
-                      Language identification at segment-level based on Heliport: (
-                      <a
-                        href="https://github.com/ZJaume/heliport"
-                        target="_blank"
-                        className={styles.tooltipLink}
-                      >
-                        https://github.com/ZJaume/heliport
-                      </a>
-                      )
+                      <p style={{ fontSize: "14px" }}>   Language identification at segment-level based on Heliport: (
+                        <a
+                          href="https://github.com/ZJaume/heliport"
+                          target="_blank"
+                          className={styles.tooltipLink}
+                        >
+                          https://github.com/ZJaume/heliport
+                        </a>
+                        )</p>
                     </Tooltip>
                   </>
                 </div>
@@ -605,21 +609,23 @@ export default function Report({ date, report }) {
                   : "Source segment length distribution by token"}
               </h3>
               <a className="segment-length-info">
-                {!footNote && <InfoCircle />}
+                {!footNote && <InfoCircle style={{ marginBottom: "-3px", marginLeft: "3px" }} />}
               </a>
               <Tooltip
                 anchorSelect=".segment-length-info"
                 place="top"
                 clickable
+                style={{ fontWeight: 400, backgroundColor: "rgba(17, 21, 24, 1)", zIndex: 10000 }}
+
               >
-                Tokenized with{" "}
-                <a
-                  href="https://github.com/hplt-project/data-analytics-tool/blob/main/tokenizers-info.md"
-                  target="_blank"
-                  className={styles.tooltipLink}
-                >
-                  https://github.com/hplt-project/data-analytics-tool/blob/main/tokenizers-info.md
-                </a>
+                <p style={{ fontSize: "14px" }}> Tokenized with{" "}
+                  <a
+                    href="https://github.com/hplt-project/data-analytics-tool/blob/main/tokenizers-info.md"
+                    target="_blank"
+                    className={styles.tooltipLink}
+                  >
+                    https://github.com/hplt-project/data-analytics-tool/blob/main/tokenizers-info.md
+                  </a></p>
               </Tooltip>
             </div>
             <div className={styles.desktopNum}>